Moved customer app from Accounts app
Moved the customer app views, and url from Accounts app
