+++ customer/urls.py
@@ -2,9 +2,6 @@ from django.urls import path
 from rest_framework_simplejwt.views import TokenRefreshView
 from .views import ForgotPasswordView, VerifyOTPView, NewPasswordView,LoginView
 
-
-
-
 urlpatterns = [
     path('forgot-password/', ForgotPasswordView.as_view(), name='forgot-password'),
     path('verify-otp/', VerifyOTPView.as_view(), name='verify-otp'),
