+++ figma/urls.py
@@ -20,7 +20,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('accounts/', include('Accounts.urls')),
+    path('user/', include('customer.urls')),
 ]
